requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai==1.54.3
python-dotenv==1.0.0
PyMuPDF==1.24.0
pytesseract==0.3.10
Pillow==10.4.0